+++ riemann/in_memory_database_test.go
@@ -11,13 +11,13 @@ import (
 var _ = Describe("In Memory Database Tests", func() {
 
 	It("is initially empty", func() {
-		var db = riemann.DivisorDb(riemann.InMemoryDivisorDb{Data: make(map[int]riemann.RiemannDivisorSum)})
+		var db = riemann.DivisorDb(riemann.InMemoryDivisorDb{Data: make(map[int64]riemann.RiemannDivisorSum)})
 		loadedData := db.Load()
 		Expect(len(loadedData)).To(Equal(0))
 	})
 
 	It("Upserts correctly", func() {
-		var db = riemann.DivisorDb(riemann.InMemoryDivisorDb{Data: make(map[int]riemann.RiemannDivisorSum)})
+		var db = riemann.DivisorDb(riemann.InMemoryDivisorDb{Data: make(map[int64]riemann.RiemannDivisorSum)})
 		records := []riemann.RiemannDivisorSum{
 			{N: 1, DivisorSum: 1, WitnessValue: 1},
 			{N: 2, DivisorSum: 2, WitnessValue: 2},
@@ -65,7 +65,7 @@ var _ = Describe("In Memory Database Tests", func() {
 	})
 
 	It("Summarizes", func() {
-		var db = riemann.DivisorDb(riemann.InMemoryDivisorDb{Data: make(map[int]riemann.RiemannDivisorSum)})
+		var db = riemann.DivisorDb(riemann.InMemoryDivisorDb{Data: make(map[int64]riemann.RiemannDivisorSum)})
 		By("correctly summarizing empty data", func() {
 			summaryData := db.Summarize()
 			expectedSummaryData := riemann.SummaryStats{
